feat: :construction: Entities update

+++ src/main/java/com/springproject/springproject/domain/services/Pais/PaisServiceImpl.java
@@ -3,6 +3,7 @@ package com.springproject.springproject.domain.services.Pais;
 import com.springproject.springproject.persistence.entities.Pais;
 import com.springproject.springproject.persistence.repositories.PaisRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -32,9 +33,10 @@ public class PaisServiceImpl implements PaisService {
     }
 
     @Override
+    @Transactional
     public Pais update(Long id, Pais pais) {
         if (paisRepository.existsById(id)) {
-            pais.setId(id);
+            pais.setIdPais(id);
             return paisRepository.save(pais);
         } else {
             throw new RuntimeException("Pais no encontrado con id " + id);
@@ -42,8 +44,14 @@ public class PaisServiceImpl implements PaisService {
     }
 
     @Override
+    @Transactional
     public Optional<Pais> delete(Long id) {
-        paisRepository.deleteById(id);
-        return null;
+        Optional<Pais> pais = paisRepository.findById(id);
+        if (pais.isPresent()) {
+            paisRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Pais no encontrado con id " + id);
+        }
+        return pais;
     }
 }
